Transport cleanup
